+++ src/Backend/Renderer/SDL_Renderer/BackendTextureSDL.cpp
@@ -111,7 +111,7 @@ namespace tgui
 
         // Copy the pixels to the texture
         if (m_texture && pixels)
-            SDL_UpdateTexture(m_texture, nullptr, pixels, size.x * 4);
+            SDL_UpdateTexture(m_texture, nullptr, pixels, static_cast<int>(size.x * 4));
 
         return (m_texture != nullptr);
     }
